requirements.txt
discord.py
aiosqlite
python-dotenv
Pillow
aiohttp
Babel
pytz